Remove **options from listen_for method
Change signature of 'listen_for' method by deleting **options
signature from it, as it was broken anyway - you had had no
option to run this with additional named options, but it was
needed to have an ability of pushing options of decorated
function. Instead of it some 'academic' code was added - now
'listen_for' method is a decorator factory which returns
decorator that have new function declared inside. It allows to
push named options of decorated function to it and get rid of
wrong signature of factory.